+++ packages/composites/proof-of-reserves/src/utils/protocol.ts
@@ -9,12 +9,14 @@ import { makeRequestFactory, callAdapter } from '.'
 
 // protocol adapters
 import * as swellList from '@chainlink/swell-address-list-adapter'
+import * as staderList from '@chainlink/stader-address-list-adapter'
 import * as renVM from '@chainlink/renvm-address-set-adapter'
 import * as wBTC from '@chainlink/wbtc-address-set-adapter'
 import * as Gemini from '@chainlink/gemini-adapter'
 import * as celsiusAddressList from '@chainlink/celsius-address-list-adapter'
 import * as chainReserveWallets from '@chainlink/chain-reserve-wallet-adapter'
 import * as wrapped from '@chainlink/wrapped-adapter'
+import * as porAddressList from '@chainlink/por-address-list-adapter'
 
 export const LIST_ADAPTER = 'LIST'
 
@@ -27,6 +29,8 @@ export const adapters: AdapterImplementation[] = [
   chainReserveWallets as unknown as AdapterImplementation,
   wrapped as unknown as AdapterImplementation,
   swellList as unknown as AdapterImplementation,
+  staderList as unknown as AdapterImplementation,
+  porAddressList as unknown as AdapterImplementation,
 ]
 
 export type Protocol = typeof adapters[number]['NAME']