+++ packages/core/legos/src/sources.ts
@@ -30,6 +30,7 @@ import * as coinapi from '@chainlink/coinapi-adapter'
 import * as coinbase from '@chainlink/coinbase-adapter'
 import * as coincodex from '@chainlink/coincodex-adapter'
 import * as coingecko from '@chainlink/coingecko-adapter'
+import * as coingecko_test from '@chainlink/coingecko-test-adapter'
 import * as coinlore from '@chainlink/coinlore-adapter'
 import * as coinmarketcap from '@chainlink/coinmarketcap-adapter'
 import * as coinmetrics from '@chainlink/coinmetrics-adapter'
@@ -112,6 +113,7 @@ import * as sochain from '@chainlink/sochain-adapter'
 import * as solana_view_function from '@chainlink/solana-view-function-adapter'
 import * as spectral_macro_score from '@chainlink/spectral-macro-score-adapter'
 import * as sportsdataio from '@chainlink/sportsdataio-adapter'
+import * as stader_address_list from '@chainlink/stader-address-list-adapter'
 import * as stader_labs from '@chainlink/stader-labs-adapter'
 import * as stasis from '@chainlink/stasis-adapter'
 import * as swell_address_list from '@chainlink/swell-address-list-adapter'
@@ -168,6 +170,7 @@ export default {
   coinbase,
   coincodex,
   coingecko,
+  coingecko_test,
   coinlore,
   coinmarketcap,
   coinmetrics,
@@ -250,6 +253,7 @@ export default {
   solana_view_function,
   spectral_macro_score,
   sportsdataio,
+  stader_address_list,
   stader_labs,
   stasis,
   swell_address_list,