+++ packages/sources/layer2-sequencer-health/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @chainlink/layer2-sequencer-health-adapter
 
+## 2.2.0
+
+### Minor Changes
+
+- 688e3a8ae: Reverting code
+
+## 2.1.0
+
+### Minor Changes
+
+- e60cb853e: add changes to support fetching Starkware Sequencer health
+
 ## 2.0.1
 
 ### Patch Changes